# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UMKM E-Commerce API",
        "description": "Dokumentasi API untuk sistem UMKM E-Commerce. Login dulu di POST /api/login, copy token dari response, lalu klik Authorize dan masukkan: Bearer {token}",
        "contact": {
            "email": "admin@toko.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Ambil semua tag",
                "operationId": "3c25208571e9aa9f69ff70ab0b7e65ec",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Buat tag baru (Admin & Manager)",
                "operationId": "bc7e5b56edd5a05e12a774c63fd9574f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Terlaris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag berhasil dibuat"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Ambil tag berdasarkan ID",
                "operationId": "52975ecdec1eccf261248d53cf2df4ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "404": {
                        "description": "Tag tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Hapus tag (Admin only)",
                "operationId": "5461d9242958e722c4bb0fca4c877fb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 6
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Tag tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{productId}/tag/{tagId}": {
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Pasang tag ke produk (Admin & Manager)",
                "operationId": "53e9fb757bb5a201c93099b5f3578ec8",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "tagId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag berhasil dipasang"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Produk atau tag tidak ditemukan"
                    },
                    "409": {
                        "description": "Tag sudah terpasang"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Lepas tag dari produk (Admin only)",
                "operationId": "d0dabf46fd42b66034867a6e298bcf24",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "tagId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag berhasil dilepas"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Produk atau tag tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token. Format: Bearer eyJ0eXAiOiJKV1Qi...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}